REF_CODE_REFACTOR: activations.cc
Refactoring the reference code for activations operators.

BUG=refactoring existing code.
