+++ tensorflow/lite/micro/kernels/BUILD
@@ -112,6 +112,7 @@ cc_library(
     name = "micro_ops",
     srcs = [
         "activations.cc",
+        "activations_common.cc",
         "add.cc",
         "add_n.cc",
         "arg_min_max.cc",
@@ -184,6 +185,7 @@ cc_library(
         "zeros_like.cc",
     ],
     hdrs = [
+        "activations.h",
         "conv.h",
         "depthwise_conv.h",
         "ethosu.h",
